@@ -52,19 +52,18 @@
     "husky": "^3.0.0",
     "jest": "^24.8.0",
     "lint-staged": "^8.2.1",
-    "nock": "^10.0.6",
+    "nock": "^12.0.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.18.2",
+    "prettier": "^2.0.4",
     "rimraf": "^2.6.3",
     "standard-version": "^6.0.1",
     "ts-jest": "^24.0.2",
     "typedoc": "^0.14.2",
     "typedoc-plugin-external-module-name": "^2.0.0",
-    "typescript": "^3.4.2"
+    "typescript": "^3.8.3"
   },
   "dependencies": {
     "@types/form-data": "^2.5.0",
-    "@types/got": "^9.4.1",
     "@types/mime-types": "^2.1.0",
     "@types/node": "^11.13.2",
     "@types/recursive-readdir": "^2.2.0",
@@ -72,7 +71,7 @@
     "fast-redact": "^1.5.0",
     "file-type": "^10.10.0",
     "form-data": "^2.5.0",
-    "got": "^9.6.0",
+    "got": "^11.0.1",
     "mime-types": "^2.1.22",
     "recursive-readdir": "^2.2.2",
     "type-fest": "^0.3.0",
