@@ -15,6 +15,7 @@ import { sleep } from '../utils/sleep';
 import { getPaginatedResource } from './utils/pagination';
 
 import events = require('events');
+import { ClientApiError } from '../utils/error';
 
 const log = debug('twilio-serverless-api:builds');
 
@@ -32,7 +33,7 @@ export async function getBuild(
   serviceSid: string,
   client: GotClient
 ): Promise<BuildResource> {
-  const resp = await client.get(`/Services/${serviceSid}/Builds/${buildSid}`);
+  const resp = await client.get(`Services/${serviceSid}/Builds/${buildSid}`);
   return (resp.body as unknown) as BuildResource;
 }
 
@@ -53,7 +54,7 @@ async function getBuildStatus(
     const resp = await getBuild(buildSid, serviceSid, client);
     return resp.status;
   } catch (err) {
-    log('%O', err);
+    log('%O', new ClientApiError(err));
     throw err;
   }
 }
@@ -72,7 +73,7 @@ export async function listBuilds(
 ): Promise<BuildResource[]> {
   return getPaginatedResource<BuildList, BuildResource>(
     client,
-    `/Services/${serviceSid}/Builds`
+    `Services/${serviceSid}/Builds`
   );
 }
 
@@ -107,17 +108,16 @@ export async function triggerBuild(
       body.AssetVersions = assetVersions;
     }
 
-    const resp = await client.post(`/Services/${serviceSid}/Builds`, {
-      // @ts-ignore
-      json: false,
+    const resp = await client.post(`Services/${serviceSid}/Builds`, {
+      responseType: 'json',
       headers: {
         'Content-Type': 'application/x-www-form-urlencoded',
       },
       body: querystring.stringify(body),
     });
-    return JSON.parse(resp.body) as BuildResource;
+    return resp.body as BuildResource;
   } catch (err) {
-    log('%O', err);
+    log('%O', new ClientApiError(err));
     throw err;
   }
 }
@@ -196,17 +196,16 @@ export async function activateBuild(
 ): Promise<any> {
   try {
     const resp = await client.post(
-      `/Services/${serviceSid}/Environments/${environmentSid}/Deployments`,
+      `Services/${serviceSid}/Environments/${environmentSid}/Deployments`,
       {
-        form: true,
-        body: {
+        form: {
           BuildSid: buildSid,
         },
       }
     );
     return resp.body;
   } catch (err) {
-    log('%O', err);
+    log('%O', new ClientApiError(err));
     throw err;
   }
 }
